fix: move compute_current_amp and adjust_precision to stableswap.rs
