+++ src/implementations/astroport/stableswap.rs
@@ -1,6 +1,9 @@
+use std::cmp::Ordering;
+
 use astroport_core::querier::{query_supply, query_token_precision};
 use astroport_core::U256;
 use astroport_pair_stable::math::compute_d;
+use astroport_pair_stable::state::{Config, CONFIG};
 use cosmwasm_schema::cw_serde;
 use cosmwasm_std::{Addr, Decimal, Env, Response, StdError, StdResult};
 use cosmwasm_std::{Deps, Uint128};
@@ -13,8 +16,7 @@ use astroport_core::asset::Asset as AstroAsset;
 use astroport_core::asset::AssetInfo as AstroAssetInfo;
 
 use super::base_pool::AstroportBasePool;
-use super::helpers::{adjust_precision, compute_current_amp, AstroAssetList};
-use super::querier::query_pair_config;
+use super::helpers::AstroAssetList;
 
 pub(crate) const N_COINS: u8 = 2;
 
@@ -52,7 +54,7 @@ impl Pool for AstroportStableSwapPool {
         assets: AssetList,
     ) -> Result<Asset, CwDexError> {
         let assets: AstroAssetList = assets.to_owned().try_into()?;
-        let config = query_pair_config(&deps.querier, &self.0.pair_addr)?;
+        let config = CONFIG.query(&deps.querier, self.0.pair_addr.clone())?;
         let mut pools: [AstroAsset; 2] =
             config.pair_info.query_pools(&deps.querier, self.0.pair_addr.to_owned())?;
         let deposits: [Uint128; 2] = [
@@ -162,4 +164,64 @@ impl Pool for AstroportStableSwapPool {
         };
         Ok(lp_token)
     }
+}
+
+/// ## Description
+/// Compute the current pool amplification coefficient (AMP).
+/// ## Params
+/// * **config** is an object of type [`Config`].
+///
+/// * **env** is an object of type [`Env`].
+///
+/// This function is needed to calculate how many LP shares a user should get when providing liquidity but is
+/// not publicly exposed in the package. Copied from the astro implementation here:
+/// https://github.com/astroport-fi/astroport-core/blob/c216ecd4f350113316be44d06a95569f451ac681/contracts/pair_stable/src/contract.rs#L1492-L1515
+fn compute_current_amp(config: &Config, env: &Env) -> StdResult<u64> {
+    let block_time = env.block.time.seconds();
+
+    if block_time < config.next_amp_time {
+        let elapsed_time =
+            Uint128::from(block_time).checked_sub(Uint128::from(config.init_amp_time))?;
+        let time_range =
+            Uint128::from(config.next_amp_time).checked_sub(Uint128::from(config.init_amp_time))?;
+        let init_amp = Uint128::from(config.init_amp);
+        let next_amp = Uint128::from(config.next_amp);
+
+        if config.next_amp > config.init_amp {
+            let amp_range = next_amp - init_amp;
+            let res = init_amp + (amp_range * elapsed_time).checked_div(time_range)?;
+            Ok(res.u128() as u64)
+        } else {
+            let amp_range = init_amp - next_amp;
+            let res = init_amp - (amp_range * elapsed_time).checked_div(time_range)?;
+            Ok(res.u128() as u64)
+        }
+    } else {
+        Ok(config.next_amp)
+    }
+}
+
+/// ## Description
+/// Return a value using a newly specified precision.
+/// ## Params
+/// * **value** is an object of type [`Uint128`]. This is the value that will have its precision adjusted.
+///
+/// * **current_precision** is an object of type [`u8`]. This is the `value`'s current precision
+///
+/// * **new_precision** is an object of type [`u8`]. This is the new precision to use when returning the `value`.
+///
+/// Copied from the astro code here:
+/// https://github.com/astroport-fi/astroport-core/blob/c216ecd4f350113316be44d06a95569f451ac681/contracts/pair_stable/src/contract.rs#L1269
+fn adjust_precision(
+    value: Uint128,
+    current_precision: u8,
+    new_precision: u8,
+) -> StdResult<Uint128> {
+    Ok(match current_precision.cmp(&new_precision) {
+        Ordering::Equal => value,
+        Ordering::Less => value
+            .checked_mul(Uint128::new(10_u128.pow((new_precision - current_precision) as u32)))?,
+        Ordering::Greater => value
+            .checked_div(Uint128::new(10_u128.pow((current_precision - new_precision) as u32)))?,
+    })
 }
